@@ -1,10 +1,13 @@
 apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
+apply plugin: 'org.jetbrains.kotlin.android'
+apply plugin: 'org.jetbrains.kotlin.plugin.compose'
 apply plugin: 'de.undercouch.download'
 apply plugin: 'maven-publish'
 
+
 android {
-    compileSdkVersion 34
+    namespace "com.assentify.sdk"
+    compileSdkVersion 36
     defaultConfig {
       applicationId "com.assentify.sdk"
 
